Ignore sink events without a valid SourceID
No longer consider events with an empty SourceID field in their metadata
as invalid. This case could only be triggered if a Google calendar event
contains `ExtendedProperties` in which the `SourceID` is set to an empty
string. However, this can not be the case for events synced with
somewhat recent CalendarSync versions.